cleaned up using routes, added use strict(protected reserved word so didn't add to mwFuncs
