+++ index.js
@@ -1,94 +1,10 @@
+"use strict";
 const express = require("express");
-// const cores = require("bcryptjs");
-const bcrypt = require("bcryptjs");
-const jwt = require("jsonwebtoken");
-const db = require("./db/dbConfig.js");
-
 const server = express();
 const PORT = 4000;
 
-server.use(express.json());
-
-server.get("/", (req, res) => {
-  res.status(201).send("Working");
-});
-// generate token function
-const secret = "rosebud";
-const generateToken = user => {
-  const payload = {
-    username: user.username,
-  };
-  const options = {
-    expiresIn: "1h",
-    jwtid: "12345",
-  };
-  return jwt.sign(payload, secret, options);
-};
-// protected middleware
-const protected = (req, res, next) => {
-  const token = req.headers.authorization;
-  if (token) {
-    jwt.verify(token, secret, (err, decodedToken) => {
-      if (err) {
-        // wrong token
-        res.status(401).json({ message: "Invalid Token" });
-      } else {
-        // good token
-        req.user = { username: decodedToken.username };
-        next();
-      }
-    });
-  } else {
-    res.status(401).json({ message: "Missing Token" });
-  }
-};
-// end middleware
-
-// register (add user)
-server.post("/api/register", (req, res) => {
-  const creds = req.body;
-  const hash = bcrypt.hashSync(creds.password, 10);
-  creds.password = hash;
-  db("users")
-    .insert(creds)
-    .then(ids => {
-      const id = ids[0];
-      db("users")
-        .where({ id })
-        .first()
-        .then(user => {
-          const token = generateToken(user);
-          res.status(201).json({ id: user.id, token });
-        })
-        .catch(err => res.status(500).send(err));
-    });
-});
-// login
-server.post("/api/login", (req, res) => {
-  const creds = req.body;
-  db("users")
-    .where({ username: creds.username })
-    .first()
-    .then(user => {
-      if (user && bcrypt.compareSync(creds.password, user.password)) {
-        const token = generateToken(user);
-        res.status(200).json({ token });
-      } else {
-        res.status(401).json({ message: "You shall not pass!" });
-      }
-    })
-    .catch(err => res.status(500).send(err));
-});
-
-// get users list
-server.get("/api/users", protected, (req, res) => {
-  db("users")
-    .select("id", "username", "password")
-    .then(users => {
-      res.json(users);
-    })
-    .catch(err => res.send(err));
-});
+const configureMiddleware = require("./middleware/middleware.js");
+configureMiddleware(server);
 
 server.listen(PORT, () => {
   console.log(`Server listening on ${PORT}`);
